+++ src/main/java/brotherdetjr/gama/ShiftFilterParams.java
@@ -6,10 +6,11 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public final class ShiftFilterParams {
     public static final String FILTER_NAME = "shift";
     private final String direction;
-    private final int distancePx;
+    private final double distancePx;
 
     @JsonCreator
-    public ShiftFilterParams(@JsonProperty("direction") String direction, @JsonProperty("distancePx") int distancePx) {
+    public ShiftFilterParams(@JsonProperty("direction") String direction,
+                             @JsonProperty("distancePx") double distancePx) {
         Direction.valueOf(direction.toUpperCase()); // simple validation
         this.direction = direction;
         this.distancePx = distancePx;
@@ -19,7 +20,7 @@ public final class ShiftFilterParams {
         return direction;
     }
 
-    public int getDistancePx() {
+    public double getDistancePx() {
         return distancePx;
     }
 }
